+++ packages/moengage_flutter/moengage_flutter_platform_interface/lib/src/internal/method_channel_moengage_flutter.dart
@@ -285,4 +285,9 @@ class MethodChannelMoEngageFlutter extends MoEngageFlutterPlatform {
     _methodChannel.invokeMethod(methodNameShowNudge,
         jsonEncode(getShowNudgeJsonPayload(position, appId)));
   }
+
+  @override
+  void registerForProvisionalPush() {
+    _methodChannel.invokeMethod(methodiOSRegisterProvisionalPush);
+  }
 }