+++ packages/moengage_geofence/moengage_geofence_android/CHANGELOG.md
@@ -1,5 +1,17 @@
 # MoEngage Geofence Android Plugin
 
+# 03-10-2024
+
+## 3.1.1
+- Fixed dependency version incompatibility issue with `moengage_flutter`
+
+# 30-09-2024
+
+## 3.1.0
+- Bundling `geofence` dependency with the plugin
+- `geofence` version updated to `4.2.0`
+- **Breaking Change**: Developers should remove the `com.moengage:geofence` dependency from the `build.gradle` as it is now included with the plugin.
+
 # 07-08-2024
 
 ## 3.0.0